+++ tunnel-server2/src/main/java/com/alibaba/arthas/tunnel/server2/config/package-info.java
@@ -1,1 +1,0 @@
-package com.alibaba.arthas.tunnel.server2.config;