+++ CHANGELOG.md
@@ -4,6 +4,6 @@
 
 ### FEATURES
 
-- 新增 `arthas-tunnel-server2` 作为 `arthas-tunnel-server` 的增强版。
-- 支持 `arthas-tunnel-server2` 自动发现 Arthas Agent 应用。
-- 支持 `arthas-tunnel-server2` 访问权限动态控制。
+- 新增 `arthas-tunnel-console` 作为 `arthas-tunnel-server` 的增强版。
+- 支持 `arthas-tunnel-console` 自动发现 Arthas Agent 应用。
+- 支持 `arthas-tunnel-console` 访问权限动态控制。